+++ src/types.ts
@@ -1,1 +1,1 @@
-type AcornNode<T = any> = import('rollup').AcornNode & Record<string, T>
+type AcornNode<T = any> = import('acorn').Node & Record<string, T>